wikijs-extractor/pyproject.toml
[tool.poetry]
name = "wikijs-extractor"
version = "0.1.0"
description = ""
authors = ["Piotr Krawiec <piotr.krawiec23@gmail.com>"]
readme = "README.md"
packages = [{include = "wikijs_extractor"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
aiohttp = {extras = ["speedups"], version = "^3.8.6"}
pydantic = "^2.4.2"
langchain = "^0.0.327"
openai = "^0.28.1"
chromadb = "^0.4.15"
langchainhub = "^0.1.13"
tiktoken = "^0.5.1"
openpyxl = "^3.1.2"
streamlit = "^1.28.0"

[tool.poetry.scripts]
wikijs-save-documents = "wikijs_extractor.loaders:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
